# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Energy Monitor API",
        "version": "1.0"
    },
    "paths": {
        "/api/config/command/{identifier}": {
            "get": {
                "tags": [
                    "commands"
                ],
                "summary": "Get a command by ID or name",
                "description": "Retrieves a specific command by its ID or name.",
                "operationId": "getCommandByIdentifier",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "ID or name of the command",
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of commands",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {}
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/config/command": {
            "post": {
                "tags": [
                    "commands"
                ],
                "summary": "Create a new command in Emca Monitor",
                "description": "Creates a new command with the provided parameters.",
                "operationId": "createCommand",
                "requestBody": {
                    "description": "command details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "command_line",
                                    "command_name",
                                    "file_id"
                                ],
                                "properties": {
                                    "command_name": {
                                        "type": "string",
                                        "example": "notify-host-by-email"
                                    },
                                    "command_line": {
                                        "type": "string",
                                        "example": "$USER1$/check_ftp -H $HOSTADDRESS$ $ARG1$"
                                    },
                                    "file_id": {
                                        "type": "string",
                                        "example": "/opt/naemon/etc/naemon/conf.d/commands.cfg"
                                    },
                                    "register": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "command created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "command created successfully"
                                        },
                                        "command": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "command_description": [
                                                    "The command_description field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
        "/api/config/command/{id}": {
            "put": {
                "tags": [
                    "commands"
                ],
                "summary": "Update a command in Emca Monitor",
                "description": "Updates an existing command with the provided parameters.",
                "operationId": "updateCommand",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the command to be updated",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "command details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "command_line",
                                    "command_name",
                                    "file_id"
                                ],
                                "properties": {
                                    "command_name": {
                                        "type": "string",
                                        "example": "notify-host-by-email"
                                    },
                                    "command_line": {
                                        "type": "string",
                                        "example": "$USER1$/check_ftp -H $HOSTADDRESS$ $ARG1$"
                                    },
                                    "file_id": {
                                        "type": "string",
                                        "example": "/opt/naemon/etc/naemon/conf.d/commands.cfg"
                                    },
                                    "register": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "command updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "command updated successfully"
                                        },
                                        "command": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "command not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "command not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "commands"
                ],
                "summary": "Delete a command in op5 Monitor",
                "description": "Deletes an existing command.",
                "operationId": "deleteCommand",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the command to be deleted",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "command deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "command deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "command not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "command not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basicAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "basic"
            }
        }
    }
}